requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
python-Levenshtein==0.12.0
requests==2.22.0
urllib3==1.25.3
